pyproject.toml
[project]
name = "Agentic-RAG"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
]
